+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## UNRELEASED
 
 **Breaking Changes**
-- In `StencilShape`, the `allowEval` and `defaultLocalized` static properties have been removed. Configure these properties using `StencilShapeGlobalSettings`. 
+- In `StencilShape`, the `allowEval` and `defaultLocalized` static properties have been removed. Configure these properties using `StencilShapeConfig`. 
 
 ## 0.10.2